Add dependencies
The main thing we need is the 'name.dmaus.schxslt:java'.
That's what we use for Schematron validation.

It has the following compile time classpath:

    compileClasspath
    \--- name.dmaus.schxslt:java:3.1.1
         +--- org.xmlresolver:xmlresolver:2.1.1
         \--- net.jcip:jcip-annotations:1.0

But the runtime classpath is more interesting:

    runtimeClasspath
    \--- name.dmaus.schxslt:java:3.1.1
         +--- org.xmlresolver:xmlresolver:2.1.1
         |    +--- org.apache.httpcomponents:httpclient:4.5.5
         |    |    +--- org.apache.httpcomponents:httpcore:4.4.9
         |    |    +--- commons-logging:commons-logging:1.2
         |    |    \--- commons-codec:commons-codec:1.10
         |    +--- org.apache.httpcomponents:httpcore:4.4.9
         |    +--- xml-apis:xml-apis:1.4.01
         |    \--- org.slf4j:slf4j-api:1.7.25
         +--- name.dmaus.schxslt:schxslt:1.8.6
         +--- net.jcip:jcip-annotations:1.0
         \--- net.sf.saxon:Saxon-HE:10.6

Main thing to notice is that there's a dependency on SLF4J API,
which means that we, as the application, need to ship with some
SLF4J backend.

So we add slf4j-simple, but logback would do just fine, too.

Also, we might as well use SLF4J ourselves, too, since it is
already part of our dependency tree.